pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.85.0",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.1",
    "telegram>=0.0.1",
]
